pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "embodied-agent-prototype"
version = "0.1.0"
description = "Desktop embodied intelligence prototype with perception-decision-execution skeleton."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "langgraph>=0.2.0",
  "PyYAML>=6.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
